Rewriting - 3(I think its finished)
