+++ Include/parser.h
@@ -5,84 +5,19 @@
 #include <iostream>
 #include <map>
 #include <memory>
-#include <vector>
 #include <variant>
+#include <vector>
 
-#include "./utility/handle.h"
-#include "ast/tree.h"
-#include "ast/node.h"
+#include "../src/utility/handle.h"
+#include "token.h"
+#include "../src/ast/tree.h"
 class DonsusParser;
 
-// lexer
+// lexer#include "../src/ast/tree.h"
 // Token types made by the lexer
-typedef enum {
-  // Tokens
-  DONSUS_NAME,    // IDENTIFIER
-  DONSUS_END,     // mark the end of the code
-  DONSUS_NUMBER,  // 69
-  DONSUS_STRING,  // "hello world"
-  DONSUS_NEWLINE, // \n
-
-  DONSUS_LPAR,          // (
-  DONSUS_RPAR,          // )
-  DONSUS_LSQB,          // [
-  DONSUS_RSQB,          // ]
-  DONSUS_COLO,          // :
-  DONSUS_COMM,          // ,
-  DONSUS_SEMICOLON,     // ;
-  DONSUS_PLUS,          // +
-  DONSUS_PLUS_EQUAL,    // +=
-  DONSUS_MINUS,         // -
-  DONSUS_MINUS_EQUAL,   // -=
-  DONSUS_STAR,          // *
-  DONSUS_STAR_EQUAL,    // *=
-  DONSUS_SLASH,         // /
-  DONSUS_SLASH_EQUAL,   // /=
-  DONSUS_LESS,          // x < y
-  DONSUS_LESS_EQUAL,    // x <= y
-  DONSUS_GREATER,       // x > y
-  DONSUS_GREATER_EQUAL, // x >= y
-  DONSUS_EQUAL,         // =
-  DONSUS_DOUBLE_EQUAL,  // ==
-  DONSUS_DOT,           // .
-  DONSUS_PERCENT,       // %
-  DONSUS_LBRACE,        // {
-  DONSUS_RBRACE,        // }
-  DONSUS_CIRCUMFLEX,    // ^
-  DONSUS_INCREMENT,     // ++
-  DONSUS_DECREMENT,     // --
-  DONSUS_EXCLAMATION,   // !
-  DONSUS_COMMENT,       // #
-  DONSUS_SINGLE_QUOTE,  // '
-  DONSUS_DOUBLE_QUOTE,  // "
-  DONSUS_THREE_DOTS,    // ...
-  DONSUS_NULL_VALUE,
-
-  // INT
-  DONSUS_BASIC_INT, // int
-  DONSUS_I8,        // [FROM -(2^7) TO (2^7) -1]
-  DONSUS_I16,       // [FROM −(2^15) TO 2^15 − 1]
-  DONSUS_I32,       // [FROM 0 TO 2^32 -1]
-  DONSUS_I64,       // [FROM 0 TO 2^64 − 1]
-  DONSUS_U32,       // [FROM 0 TO 2^32 -1]
-  DONSUS_U64,       // [FROM 0 TO 2^64 − 1]
-
-  DONSUS_BOOL, // bool
-  DONSUS_VOID, // void
-  DONSUS_CHAR, // 'D'
-} donsus_token_kind;
 
 extern std::map<std::string, donsus_token_kind> DONSUS_TYPES_LEXER;
 
-struct donsus_token {
-  donsus_token_kind kind;  // the kind of the token
-  std::string value;       // the value of the token(string)
-  unsigned int length;     // the length of the token
-  unsigned int line;       // the line number of the token(starts from one)
-  unsigned int precedence; // precedence of the token(the higher the value the
-                           // more precedence the token has)
-
-};
 // lexer--
 struct donsus_lexer {
   donsus_lexer(std::string input)
@@ -103,20 +38,21 @@ public:
   using parse_result = utility::handle<donsus_ast::node>;
   DonsusParser(donsus_lexer &lexer);
   donsus_token donsus_parser_next();
-  auto donsus_parse() -> parse_result ;
+  auto donsus_parse() -> parse_result;
   void print_token();
   donsus_token donsus_peek();
 
   // create node
-  auto create_node(donsus_ast::donsus_node_type type, uint64_t child_count, donsus_token value) -> utility::handle<donsus_ast::node>{
+  auto create_node(donsus_ast::donsus_node_type type, uint64_t child_count,
+                   donsus_token value) -> utility::handle<donsus_ast::node> {
     return donsus_tree.create_node(type, child_count, value);
   }
   // parsing expression
   auto donsus_expr() -> parse_result;
   // parsing number expressions
   auto donsus_number_expr(unsigned int ptp) -> parse_result;
-  auto donsus_number(donsus_token_kind type) -> parse_result;
-  auto donsus_number_primary() -> parse_result;
+  auto donsus_number_primary(donsus_ast::donsus_node_type type, uint64_t child_count)
+      -> parse_result;
 
   // parsing variable declaration
   auto donsus_variable_decl(donsus_token_kind type) -> parse_result;
