+++ Include/sema.h
@@ -1,10 +1,13 @@
 #ifndef DONSUS_SEMA_H
 #define DONSUS_SEMA_H
 #include "../Include/symbol_table.h"
+#include "../src/utility/handle.h"
+#include "../src/utility/memory_alloc.h"
 #include "parser.h"
 
-donsus_global_ast*
-donsus_sema(donsus_global_ast* ast, std::string &file_name);
-donsus_symtable* donsus_sym(std::string &file_name);
+DonsusParser::parse_result
+donsus_sema(DonsusParser::parse_result ast, std::string &file_name);
+
+utility::handle<donsus_symtable> donsus_sym(std::string &file_name, utility::DonsusAllocator& alloc);
 
 #endif // DONSUS_SEMA_H
