+++ src/ast/node.h
@@ -1,24 +1,30 @@
 #ifndef donsus_node_h
 #define donsus_node_h
-#include "../utility/handle.h"
-#include "../utility/slices.h"
-#include "../../Include/parser.h" // donsus_token as value
-// Contains ast(node) types that the parser figures out
 #include <iostream>
 #include <vector>
 
+#include "../utility/handle.h"
+#include "../utility/slices.h"
+#include "../../Include/token.h"
+
+// Contains ast(node) types that the parser figures out
+
 namespace donsus_ast {
 struct donsus_node_type {
-  enum underlying : int { DONSUS_VARIABLE_DECLARATION, DONSUS_NUMBER_EXPRESSION};
+  enum underlying : int {
+    DONSUS_VARIABLE_DECLARATION,
+    DONSUS_NUMBER_EXPRESSION
+  };
+
   donsus_node_type() = default;
   donsus_node_type(underlying type);
 
-  auto to_string() const -> std::string;
+  [[nodiscard]] auto to_string() const -> std::string;
   underlying type;
 };
 struct node {
   // children tbd
-  utility::slice<utility::handle<node>> children; // size type in the future
+  std::vector<utility::handle<node>> children; // size type in the future
   donsus_node_type type;
   donsus_token value;
 };
