+++ src/parser.cc
@@ -1,9 +1,7 @@
 // pratt parser
 #include "../Include/parser.h"
 #include "../Include/donsus.h"
-#include "../src/utility/handle.h"
 #include <iostream>
-#include <memory>
 
 /*
 // HELPER
@@ -54,8 +52,6 @@ auto DonsusParser::donsus_parse() -> parse_result {
   *this = save;
 #endif
 
-  donsus_global_ast *result;
-
   while (cur_token.kind != DONSUS_END) {
     donsus_parser_next();
     switch (cur_token.kind) {
@@ -69,7 +65,7 @@ auto DonsusParser::donsus_parse() -> parse_result {
     case DONSUS_I64:
     case DONSUS_U32:
     case DONSUS_U64: {
-      result->body.emplace_back(donsus_variable_decl(cur_token.kind));
+      donsus_tree.add_node(donsus_variable_decl(cur_token.kind));
     }
     default: {
     }
@@ -79,16 +75,15 @@ auto DonsusParser::donsus_parse() -> parse_result {
   // std::cout << "AST: " << "\n";
   // // print_ast(result);
   // #endif
-
-  return result;
+/*  return donsus_tree;*/
 }
 
 auto DonsusParser::donsus_number_expr(unsigned int ptp) -> parse_result {
   // Gt the integer on the left
-  const utility::handle<donsus_ast::node> left;
-  const utility::handle<donsus_ast::node> right;
+  utility::handle<donsus_ast::node> left;
+  utility::handle<donsus_ast::node> right;
 
-  left = donsus_number_primary(); // return AST node with only value
+  left = donsus_number_primary(donsus_ast::donsus_node_type::DONSUS_NUMBER_EXPRESSION, 20); // return AST node with only value
 
   donsus_token previous_token = cur_token; // SAVE CUR_TOKEN
 
@@ -103,6 +98,9 @@ auto DonsusParser::donsus_number_expr(unsigned int ptp) -> parse_result {
 
     left = create_node(donsus_ast::donsus_node_type::DONSUS_NUMBER_EXPRESSION, 2, previous_token);
 
+    left->children[0] = left;
+    left->children[1] = right;
+
     if (cur_token.kind == DONSUS_END) {
       return left;
     }
@@ -110,15 +108,11 @@ auto DonsusParser::donsus_number_expr(unsigned int ptp) -> parse_result {
   return left;
 }
 
-auto DonsusParser::donsus_number(donsus_ast::donsus_node_type type)
+auto DonsusParser::donsus_number_primary(donsus_ast::donsus_node_type type, uint64_t child_count)
     -> parse_result {
-  const utility::handle<donsus_ast::node> node = create_node();
-  node->children[0] = left;
-  node->children[1] = right
-}
-
-auto DonsusParser::donsus_number_primary() -> parse_result {
-  return donsus_number(cur_token.kind);
+  const utility::handle<donsus_ast::node> node = create_node(type, child_count, cur_token);
+  donsus_parser_next();
+  return node;
 }
 
 auto DonsusParser::donsus_expr() -> parse_result {
@@ -131,6 +125,6 @@ auto DonsusParser::donsus_variable_decl(donsus_token_kind type)
   // add this ast node to the top level AST
   // add this to the symbol table
   // figure out whether it has a definition
-  auto *n = new donsus_math_expr(cur_token, DONSUS_VAR_DECLARATION);
-  return n;
+/*  auto *n = new donsus_math_expr(cur_token, DONSUS_VAR_DECLARATION);
+  return n;*/
 }
