+++ src/sema.cc
@@ -21,16 +21,15 @@
 #include "../Include/symbol_table.h"
 
 // Make symbol table
-std::unique_ptr<donsus_symtable> donsus_sym(std::string &file_name) {
-  std::unique_ptr<donsus_symtable> t;
-  std::unique_ptr<donsus_symbol> s;
+utility::handle<donsus_symtable> donsus_sym(std::string &file_name, utility::DonsusAllocator& alloc) {
+  utility::handle<donsus_symtable> t;
 
-  t = donsus_sym_make(file_name); // construct a symbol table GLOBAL
+  t = donsus_sym_make(file_name, alloc); // construct a symbol table GLOBAL
   return t;
 }
 
-std::unique_ptr<donsus_global_ast>
-donsus_sema(std::unique_ptr<donsus_global_ast> ast, std::string &file_name) {
+DonsusParser::parse_result
+donsus_sema(DonsusParser::parse_result ast, std::string &file_name) {
   // TBD:
   // make sema
   // donsus_sym(file_name); // ahh, might use it here instead of runner.cc
